Teach Goal how to toString
To do this I made all the unit types know how to toTi and then taught them each to print in shortened units based on size.  Probably still not quite matching the game (e.g. probably will do 200x instead of 0.2y in some cases).

Tested this and in doing so found a bug in the mPa constructor.